s/log_manager: made removing ntp folder optional while removing ntp
When NTP folder is always deleted when NTP is deleted it is later
impossible to differentiate if NTP was successfully deleted or its
creation failed. In order to be able to distinguish failed creation from
successful deletion we will leave NTP directory intact after deleting
its content. This will allow cluster::controller_backend to track
partition movement progress.

+++ src/v/storage/log_manager.cc
@@ -220,32 +220,39 @@ ss::future<log> log_manager::do_manage(ntp_config cfg) {
     });
 }
 
-ss::future<> log_manager::remove(model::ntp ntp) {
+ss::future<>
+log_manager::remove(model::ntp ntp, log_manager::remove_dir should_remove_dir) {
     vlog(stlog.info, "Asked to remove: {}", ntp);
-    return ss::with_gate(_open_gate, [this, ntp = std::move(ntp)] {
-        auto handle = _logs.extract(ntp);
-        if (handle.empty()) {
-            return ss::make_ready_future<>();
-        }
-        // 'ss::shared_ptr<>' make a copy
-        storage::log lg = handle.mapped().handle;
-        vlog(stlog.info, "Removing: {}", lg);
-        // NOTE: it is ok to *not* externally synchronize the log here
-        // because remove, takes a write lock on each individual segments
-        // waiting for all of them to be closed before actually removing the
-        // underlying log. If there is a background operation like
-        // compaction or so, it will block correctly.
-        auto ntp_dir = lg.config().work_directory();
-        ss::sstring topic_dir = lg.config().topic_directory().string();
-        return lg.remove()
-          .then([dir = std::move(ntp_dir)] { return ss::remove_file(dir); })
-          .then([this, dir = std::move(topic_dir)]() mutable {
-              // We always dispatch topic directory deletion to core 0 as
-              // requests may come from different cores
-              return dispatch_topic_dir_deletion(std::move(dir));
-          })
-          .finally([lg] {});
-    });
+    return ss::with_gate(
+      _open_gate, [this, ntp = std::move(ntp), should_remove_dir] {
+          auto handle = _logs.extract(ntp);
+          if (handle.empty()) {
+              return ss::make_ready_future<>();
+          }
+          // 'ss::shared_ptr<>' make a copy
+          storage::log lg = handle.mapped().handle;
+          vlog(stlog.info, "Removing: {}", lg);
+          // NOTE: it is ok to *not* externally synchronize the log here
+          // because remove, takes a write lock on each individual segments
+          // waiting for all of them to be closed before actually removing the
+          // underlying log. If there is a background operation like
+          // compaction or so, it will block correctly.
+          auto ntp_dir = lg.config().work_directory();
+          ss::sstring topic_dir = lg.config().topic_directory().string();
+          auto f = lg.remove();
+          if (!should_remove_dir) {
+              return f.finally([lg] {});
+          }
+
+          return f
+            .then([dir = std::move(ntp_dir)] { return ss::remove_file(dir); })
+            .then([this, dir = std::move(topic_dir)]() mutable {
+                // We always dispatch topic directory deletion to core 0
+                // as requests may come from different cores
+                return dispatch_topic_dir_deletion(std::move(dir));
+            })
+            .finally([lg] {});
+      });
 }
 
 ss::future<> log_manager::dispatch_topic_dir_deletion(ss::sstring dir) {
